Move framing tests to a dedicated file
